libbinder_ndk: explicit 64 -> 32 bit conversions
Just in case. This was causing errors in some clients, and it's good to
get a compiler error for these things.

Fixes: 143556646
Test: CtsNdkBinderTestCases on TH

+++ libs/binder/ndk/ibinder.cpp
@@ -143,7 +143,7 @@ status_t ABBinder::dump(int fd, const ::android::Vector<String16>& args) {
         utf8Pointers.push_back(utf8Args[i].c_str());
     }
 
-    return onDump(this, fd, utf8Pointers.data(), utf8Pointers.size());
+    return onDump(this, fd, utf8Pointers.data(), static_cast<int32_t>(utf8Pointers.size()));
 }
 
 status_t ABBinder::onTransact(transaction_code_t code, const Parcel& data, Parcel* reply,
